+++ internal/app/wwctl/overlay/show/main_test.go
@@ -13,8 +13,11 @@ import (
 )
 
 var (
-	overlayCont = `
+	overlayEmail = `
 {{ if .Tags.email }}eMail: {{ .Tags.email }}{{else}} noMail{{- end }}
+`
+	overlayOverlay = `
+overlay name {{ .Overlay }}
 `
 )
 
@@ -37,7 +40,8 @@ nodes:
       - empty
 `)
 
-	env.WriteFile(t, path.Join(testenv.WWOverlaydir, "testoverlay/email.ww"), overlayCont)
+	env.WriteFile(t, path.Join(testenv.WWOverlaydir, "testoverlay/email.ww"), overlayEmail)
+	env.WriteFile(t, path.Join(testenv.WWOverlaydir, "testoverlay/overlay.ww"), overlayOverlay)
 	defer env.RemoveAll(t)
 	warewulfd.SetNoDaemon()
 	t.Run("overlay show raw", func(t *testing.T) {
@@ -49,7 +53,7 @@ nodes:
 		wwlog.SetLogWriter(buf)
 		err := baseCmd.Execute()
 		assert.NoError(t, err)
-		assert.Contains(t, buf.String(), overlayCont)
+		assert.Contains(t, buf.String(), overlayEmail)
 	})
 	t.Run("overlay show rendered node tag", func(t *testing.T) {
 		baseCmd.SetArgs([]string{"-r", "node1", "testoverlay", "email.ww"})
@@ -83,5 +87,16 @@ nodes:
 		err := baseCmd.Execute()
 		assert.NoError(t, err)
 		assert.Contains(t, buf.String(), "noMail")
+	})
+	t.Run("overlay shows overlay", func(t *testing.T) {
+		baseCmd.SetArgs([]string{"-r", "node1", "testoverlay", "overlay.ww"})
+		baseCmd := GetCommand()
+		buf := new(bytes.Buffer)
+		baseCmd.SetOut(buf)
+		baseCmd.SetErr(buf)
+		wwlog.SetLogWriter(buf)
+		err := baseCmd.Execute()
+		assert.NoError(t, err)
+		assert.Contains(t, buf.String(), "testoverlay")
 	})
 }