+++ internal/pkg/overlay/overlay.go
@@ -350,6 +350,18 @@ func RenderTemplateFile(fileName string, data TemplateStruct) (
 	err error) {
 	backupFile = true
 	writeFile = true
+	// parse the overlay name out of the absolute path
+	overlayPath, _ := filepath.Rel(OverlaySourceTopDir(), fileName)
+	withoutRootfs, err := filepath.Rel("rootfs", overlayPath)
+	if err != nil {
+		overlayPath = withoutRootfs
+	}
+	overlayPathElem := strings.Split(overlayPath, "/")
+	if len(overlayPathElem) > 1 {
+		data.Overlay = overlayPathElem[0]
+	} else {
+		data.Overlay = ""
+	}
 	tmpl, err := template.New(path.Base(fileName)).Option("missingkey=default").Funcs(template.FuncMap{
 		// TODO: Fix for missingkey=zero
 		"Include":      templateFileInclude,