+++ CHANGELOG.md
@@ -38,6 +38,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Document warewulf.conf:paths. #635
+- New "Overlay" template variable contains the name of the overlay being built. #1052  
 
 ### Changed